Updated files
I moved the testing the files to the test folder. The main test html
file shows the script to ask for a request from the website. There are
two javascript files that will show when run, if you take a look at your
developer tools and read the response from the network and see some
problems I ran into. I used a combination of ajax, jquery, and two other
frameworks: dojo toolkit and vue js to test out the api to get a
response.